+++ SSJD.Entities/StoreEntity/Promotion.cs
@@ -12,7 +12,7 @@ namespace SSJD.Entities.StoreEntity
         public string Name { get; set; } = "";
         public DateTime CreateDate { get; set; } = DateTime.Now;
         public DateTime EndDate { get; set; } = DateTime.Now;
-        public decimal PercentDiscount { get; set; } = 0;
+        public int PercentDiscount { get; set; } = 0;
 
         public OrderDetail OrderDetail { get; set; }
