+++ SSJD.ViewModel/StoreViewModel/MemberCard/MemberCardRequestModel.cs
@@ -11,6 +11,6 @@ namespace SSJD.ViewModel.StoreViewModel.MemberCard
         public string ID { get; set; } = "";
         public int Point { get; set; } = 0;
         public string MemberClass { get; set; } = "";
-        public decimal Discount { get; set; } = 0;
+        public int Discount { get; set; } = 0;
     }
 }
